+++ demo/tests/__init__.py
@@ -1,3 +1,6 @@
+# -*- coding: utf8 -*-
+import json
+
 from django.test import TestCase, LiveServerTestCase
 from django.core.management import call_command
 
@@ -31,7 +34,7 @@ class LoaddataLiveTestCase(LoaddataMixin, LiveServerTestCase):
 class RegistryTestGeneric(LoaddataTestCase):
 
     def _test_registry_keys(self, keys):
-        self.assertEqual(len(keys), 20)
+        self.assertEqual(len(keys), 21)
         self.assertIn("AutocompleteColor", keys)
         self.assertIn("AutocompleteColorExtra", keys)
         self.assertIn("AutocompletePerson", keys)
@@ -58,6 +61,7 @@ class RegistryTestGeneric(LoaddataTestCase):
         self.assertIn("AutocompleteUrlConvertComplex", keys)
         self.assertIn("AutocompleteUrlSimpleAuth", keys)
         self.assertIn("AutocompleteUrlHeadersAuth", keys)
+        self.assertIn("AutocompleteUrlErrors", keys)
 
 
 class MockRequestUser(object):
@@ -68,3 +72,10 @@ class MockRequestUser(object):
 
     def is_authenticated(self):
         return self._is_authenticated
+
+
+def get_json(response, key='data'):
+    data = json.loads(response.content.decode())
+    if key:
+        return data.get(key, None)
+    return data